requirements.txt
# Core
Django==5.1.7
django-configurations==2.5.1
gunicorn==23.0.0
newrelic==10.3.1
setuptools==75.6.0
djoser==2.3.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
# Async Tasks
celery==5.4.0
redis==5.0.1
django-celery-beat==2.8.1
# For the persistence stores
psycopg2-binary==2.9.10
dj-database-url==2.3.0
pillow==11.1.0
# Model Tools
django-model-utils==5.0.0
django_unique_upload==0.2.1
# Rest apis
djangorestframework==3.15.2
django-filter==24.3
# Developer Tools
ipdb==0.13.13
ipython==8.30.0
mkdocs==1.6.1
flake8==7.1.1
# Testing
mock==5.1.0
factory-boy==3.3.1
pytest-django==4.9.0
# Static and Media Storage
django-storages==1.14.4
boto3==1.35.76
drf-spectacular == 0.28.0
# Data Import/Export
django-import-export==4.3.2
pandas==2.2.0
openpyxl==3.1.2
#Admin Enhancements
django-admin-interface==0.29.4
django-colorfield==0.11.0
# Audit-log
django-simple-history==3.4.0
django-auditlog==2.3.0
# PDF Processing
reportlab==4.4.1
PyPDF2==3.0.1